pyproject.toml
[tool.poetry]
name = "pi_playbooks"
version = "0.0.1"
description = ""
authors = ["Roman Geraskin <rgeraskin@users.noreply.github.com>"]

[tool.poetry.dependencies]
# if your playbook requires additional dependencies, add them here
#some-dependency = "^1.2.3"

[tool.poetry.dev-dependencies]
robusta-cli = "^0.8.9"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"